+++ Common/src/main/java/net/potionstudios/biomeswevegone/tags/BWGBlockTags.java
@@ -1,6 +1,7 @@
 package net.potionstudios.biomeswevegone.tags;
 
 import net.minecraft.core.registries.Registries;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.tags.TagKey;
 import net.minecraft.world.level.block.Block;
 import net.potionstudios.biomeswevegone.BiomesWeveGone;
@@ -28,12 +29,16 @@ public class BWGBlockTags {
     public static final TagKey<Block> DAFFODILS = create("flowers/daffodils");
 
     /** Storage Blocks **/
-    public static final TagKey<Block> STORAGE_BLOCKS_ALLIUM = create("storage_blocks/allium");
-    public static final TagKey<Block> STORAGE_BLOCKS_PINK_ALLIUM = create("storage_blocks/pink_allium");
-    public static final TagKey<Block> STORAGE_BLOCKS_WHITE_ALLIUM = create("storage_blocks/white_allium");
-    public static final TagKey<Block> STORAGE_BLOCKS_ROSE = create("storage_blocks/rose");
+    public static final TagKey<Block> STORAGE_BLOCKS_ALLIUM = createCommon("storage_blocks/allium");
+    public static final TagKey<Block> STORAGE_BLOCKS_PINK_ALLIUM = createCommon("storage_blocks/pink_allium");
+    public static final TagKey<Block> STORAGE_BLOCKS_WHITE_ALLIUM = createCommon("storage_blocks/white_allium");
+    public static final TagKey<Block> STORAGE_BLOCKS_ROSE = createCommon("storage_blocks/rose");
 
     private static TagKey<Block> create(String name) {
         return TagKey.create(Registries.BLOCK, BiomesWeveGone.id(name));
+    }
+
+    private static TagKey<Block> createCommon(String name) {
+        return TagKey.create(Registries.BLOCK, ResourceLocation.fromNamespaceAndPath("c", name));
     }
 }